Dto 자료형 변경

+++ src/main/java/com/semi/easycoding/education/dto/EducationDto.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @Setter
 public class EducationDto {
     private Long EducationId;
-    private Integer EducationType;
+    private Short EducationType;
     private String EducationTitle;
     private String EducationContent;
     private LocalDateTime createdAt;
